clear the files to merge
