--- client/src/Footer.css
@@ -1,12 +1,0 @@
-.footer {
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-    text-align: center;
-    background-color: transparent; 
-    padding: 10px 0;
-    color: white;
-    font-family: Arial, sans-serif;
-    font-size: 14px;
-    font-weight: bold;
-}
