--- client/src/Footer.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import './Footer.css';
-
-const Footer = () => {
-  return (
-    <footer className="footer">
-      © 2023 ASEAN-Australia Strategic Youth Partnership Ltd ACN 631 871 184
-    </footer>
-  );
-};
-
-export default Footer;
