--- client/src/HomePage.css
@@ -1,30 +1,0 @@
-.infoSection {
-  flex: 1;
-  padding: 20px;
-  color: #FFFFFF;
-  text-align: left;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  max-width: 50%; /* Define a max width for the text section */
-}
-
-.description {
-  font-size: 18px;
-  font-weight: bold;
-  line-height: 1.5;
-  word-wrap: break-word; /* Ensure text wraps within its container */
-}
-
-.videoContainer {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.videoSmall {
-  width: 1%; /* Keep the video size */
-  max-width: 600px; /* Set a maximum width to maintain its original size */
-  border-radius: 10px;
-}
