--- client/src/Navbar.css
@@ -1,42 +1,0 @@
-.navbar {
-    display: flex;
-    align-items: center;
-    background-color: #2b4b2d; 
-    padding: 10px 20px;
-  }
-  
-  .navbar-logo img {
-    width: 50px; 
-  }
-  
-  .navbar-menu {
-    list-style: none;
-    display: flex;
-    flex-grow: 1;
-    justify-content: flex-end;
-    margin: 0;
-    padding: 0;
-  }
-  
-  .navbar-menu li {
-    margin: 0 20px;
-  }
-  
-  .navbar-menu li a {
-    text-decoration: none;
-    color: white;
-    font-weight: bold;
-  }
-  
-  .navbar-avatar img {
-    width: 40px; 
-    border-radius: 50%;
-  }
-  
-  .navbar-login a {
-    text-decoration: none;
-    color: white;
-    font-weight: bold;
-  }
-
-
